Refactor CPP server
