+++ game_cpp/main.cpp
@@ -1,0 +1,8 @@
+#include "Server.hpp"
+
+int main()
+{
+    Server server(1100);
+    server.start();
+    return 0;
+}
